Nano/requirements.txt
discord.py
praw
aiohttp
asyncio
isodate
ytpy
validators
pixivpy-async
asyncpraw
